Fix ddvfile
